refactor setting

+++ app/Common/Console/Commands/DB/Fursie.php
@@ -13,6 +13,9 @@ class Fursie extends Command
 
     public function handle(): void
     {
+        if (config('app.template') !== 'fursie'){
+            return;
+        }
         Schema::dropAllTables();
         Schema::disableForeignKeyConstraints();
         $db = storage_path('db/db.sql');
